[wallet/desktop] task: rename mocks file
