+++ __tests__/components/TableDisplay.spec.ts
@@ -19,8 +19,8 @@ import { getComponent } from '@MOCKS/Components';
 import { NetworkConfigurationModel } from '@/core/database/entities/NetworkConfigurationModel';
 import { simpleWallet1, simpleWallet2 } from '@MOCKS/Accounts';
 import { mockMosaicRowValue, mosaicsMock } from '@MOCKS/mosaics';
-import { mockNamespaceRowValue, mockNamespaces } from '@MOCKS/Namespaces';
-import { mockMetadatas } from '@MOCKS/Metadatas';
+import { mockNamespaceRowValue, mockNamespaces } from '@MOCKS/namespaces';
+import { mockMetadatas } from '@MOCKS/metadatas';
 import { Signer } from '@/store/Account';
 import { Address, NamespaceId, MosaicId } from 'symbol-sdk';
 import ModalMetadataDisplay from '@/views/modals/ModalMetadataDisplay/ModalMetadataDisplay.vue';
